Supplemental_Data and manuscript figure update
In this commit I clean up the repo by addring an archive directory, as well as move all of Viktoria's figures into its own directory.  I also added the latest versions of the jupyter notebooks, tables and figures for the manuscript and added a directory called supplementary data.  I also added the files that were over 100mb to the gitignore.
 Finally, I moved all the taccc script, launcher and otherwise into a direcotry called tacc scripts.